Lucia beta 12, split auth shared.ts to index.ts and oauth.ts
